+++ src/LinAlg/hiopMatrixSparseTriplet.hpp
@@ -23,7 +23,7 @@ public:
   virtual void setToConstant(double c);
   virtual void copyFrom(const hiopMatrixSparseTriplet& dm);
 
-  virtual void copyRowsFrom(const hiopMatrix& src, const int* rows_idxs, int n_rows);
+  virtual void copyRowsFrom(const hiopMatrix& src, const long long* rows_idxs, long long n_rows);
   
   virtual void timesVec(double beta,  hiopVector& y,
 			double alpha, const hiopVector& x) const;
